v5 eggnog performance improvement.
We have added a emapper bash script as a wrapper for emapper.py.

This script is replacing the emapper --data-dir / eggnog.db file
and loading that file in /dev/shm for faster access.

Running this requires having at least 40GB of RAM as /dev/shm is
using the RAM for it.